fixed up points

+++ src/displayFiles/Point.java
@@ -2,15 +2,15 @@ package displayFiles;
 
 public class Point {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int xval, int yval){
+    public Point(double xval, double yval){
         x = xval;
         y = yval;
     }
 
-    public int getX(){return x; }
-    public int getY(){return y; }
+    public double getX(){return x; }
+    public double getY(){return y; }
 
 }
